ProductTracking/requirements.txt
python-dotenv==0.15.0
python-http-client==3.3.2
sendgrid==6.6.0
starkbank-ecdsa==1.1.0
urllib3==1.26.3